refactor: Creating the utils/ folder
