+++ git_scratch/utils/read_object.py
